+++ requirements.txt
@@ -8,3 +8,4 @@ einops ==0.7.0
 efficientnet_pytorch==0.7.1
 zetascale ==0.9.9
 classifier-free-guidance-pytorch ==0.5.1
+octo @ git+https://github.com/sebbyjp/octo.git@peralta
