+++ rtx/__init__.py
@@ -1,13 +1,9 @@
 from rtx.rtx2 import RTX2
 from rtx.rtx1 import RTX1
 from rtx.efficient_net import EfficientNetFilm
-from rtx.data_util import describe, format_imgs, preprocess
 
 __all__ = [
     "RTX2",
     "RTX1",
     "EfficientNetFilm",
-    "describe",
-    "format_imgs",
-    "preprocess",
 ]
