+++ rtx/data/registry.py
@@ -1,0 +1,4 @@
+DATASET_MIXES = {
+    "oxe_magic_soup",
+    "ee_pose_magic_soup"
+}
