+++ rtx/rtx1.py
@@ -3,7 +3,7 @@ import torch
 import torch.nn.functional as F
 from torch import nn, einsum, Tensor
 from typing import List, Optional, Callable, Tuple
-# from beartype import beartype
+from beartype import beartype
 from einops import pack, unpack, repeat, reduce, rearrange
 from einops.layers.torch import Rearrange, Reduce
 
@@ -521,7 +521,7 @@ class FilmMaxVit(nn.Module):
             nn.Linear(self.embed_dim, config.num_classes, bias=False),
         )
 
-    # @beartype
+    @beartype
     def forward(
         self,
         x,
@@ -643,7 +643,7 @@ class TransformerAttention(nn.Module):
         return self.to_out(out)
 
 
-# @beartype
+@beartype
 class Transformer(nn.Module):
     def __init__(
         self,
@@ -783,7 +783,7 @@ class RT1Config:
         self.use_attn_conditioner = use_attn_conditioner
 
 
- #@beartype
+@beartype
 class RT1(nn.Module):
     def __init__(
         self,
