+++ run.py
@@ -1,13 +1,20 @@
 from examples import rtx1_example, rtx2_example, train_example
 from rtx import RTX1, RTX2
 from rtx.rtx1 import FilmViTConfig
+from rtx.action_tokenization import RTX1ActionTokenizer
 from absl import app, flags, logging
-import sys
 
 REGISTRY = {
-    "rtx1": RTX1,
-    "rtx2": RTX2,
+    "rtx1": {
+        "model": RTX1,
+        "action_tokenizer": RTX1ActionTokenizer,
+    },
+    "rtx2": {
+        "model": RTX2,
+        "action_tokenizer": None,
+    },
 }
+
 
 MODES = ["inference", "train"]
 
@@ -27,8 +34,9 @@ def main(_):
     elif FLAGS.mode == "train":
         if FLAGS.pretrained_vit and FLAGS.model == "rtx2":
             logging.fatal("Option `pretrained_vit` is not available for model {} ".format(FLAGS.model))
-        model = REGISTRY[FLAGS.model](vit_config= FilmViTConfig(pretrained= FLAGS.pretrained_vit))
-        train_example.run(model)
+        model = REGISTRY[FLAGS.model]['model'](vit_config= FilmViTConfig(pretrained= FLAGS.pretrained_vit))
+        action_tokenizer = REGISTRY[FLAGS.model]['action_tokenizer']()
+        train_example.run(model, action_tokenizer)
 
 if __name__ == "__main__":
     app.run(main)
